requirements.txt
absl-py==0.9.0
astor==0.8.1
backcall==0.1.0
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.1
chardet==3.0.4
coverage==5.1
DateTime==4.3
decorator==4.4.2
gast==0.2.2
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
imgkit==1.0.2
importlib-metadata==1.6.0
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
langdetect==1.0.8
Markdown==3.2.2
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.4
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
Pygments==2.6.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.1
regex==2020.5.14
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
tensorboard==1.15.0
tensorflow==1.15.2
tensorflow-estimator==1.15.1
termcolor==1.1.0
toposort==1.5
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
tweepy==3.8.0
urllib3==1.25.9
Wand==0.6.1
wcwidth==0.2.2
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.1.0
